initial re-structing pass
